examples/demo/requirements.txt
Django>=1.3
django-grappelli
django-haystack
django-localeurl
git+git://github.com/django-mptt/django-mptt.git#egg=django-mptt
django-images
django-pagination
django-prices
south
whoosh
# these need setup.py
git+git://github.com/mirumee/django-payments.git#egg=django-payments
# git+git://github.com/emesik/mamona.git#egg=mamona
